+++ winwin-be-api/src/test/java/com/dpm/winwin/api/MemberControllerTest.java
@@ -1,23 +1,16 @@
 package com.dpm.winwin.api;
 
 import static org.mockito.BDDMockito.given;
-import com.dpm.winwin.api.member.application.MemberCommandService;
-import com.dpm.winwin.api.member.application.MemberQueryService;
+import com.dpm.winwin.api.member.service.MemberCommandService;
 import com.dpm.winwin.api.member.dto.request.MemberCreateRequest;
 import com.dpm.winwin.api.member.dto.response.MemberCreateResponse;
 import com.dpm.winwin.api.utils.RestDocsTestSupport;
-import com.dpm.winwin.domain.repository.member.dto.response.MemberReadResponse;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.restdocs.AutoConfigureRestDocs;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
-import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.MediaType;
 import org.springframework.restdocs.payload.JsonFieldType;
-import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.ResultActions;
 
 import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.document;
